+++ central/detection/deploytime/single_deployment_executor.go
@@ -11,7 +11,6 @@ import (
 	"github.com/stackrox/rox/central/globalindex"
 	imageIndexer "github.com/stackrox/rox/central/image/index"
 	"github.com/stackrox/rox/central/searchbasedpolicies"
-	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/images/types"
 	"github.com/stackrox/rox/pkg/search"
@@ -19,11 +18,10 @@ import (
 	"github.com/stackrox/rox/pkg/utils"
 )
 
-func newSingleDeploymentExecutor(executorCtx context.Context, ctx DetectionContext, searcher search.Searcher, deployment *storage.Deployment, images []*storage.Image) alertCollectingExecutor {
+func newSingleDeploymentExecutor(executorCtx context.Context, ctx DetectionContext, deployment *storage.Deployment, images []*storage.Image) alertCollectingExecutor {
 	return &policyExecutor{
 		executorCtx: executorCtx,
 		ctx:         ctx,
-		searcher:    searcher,
 		deployment:  deployment,
 		images:      images,
 	}
@@ -36,7 +34,6 @@ type closeableIndex interface {
 type policyExecutor struct {
 	executorCtx context.Context
 	ctx         DetectionContext
-	searcher    search.Searcher
 	deployment  *storage.Deployment
 	images      []*storage.Image
 	alerts      []*storage.Alert
@@ -76,36 +73,8 @@ func (d *policyExecutor) Execute(compiled detection.CompiledPolicy) error {
 	return nil
 }
 
-func (d *policyExecutor) hasImagesWithNoIDs() bool {
-	for _, c := range d.deployment.GetContainers() {
-		if c.GetImage().GetId() == "" {
-			return true
-		}
-	}
-	return false
-}
-
 func (d *policyExecutor) getViolations(ctx context.Context, enforcement storage.EnforcementAction, matcher searchbasedpolicies.Matcher) ([]*storage.Alert_Violation, error) {
-	if d.hasImagesWithNoIDs() {
-		return matchWithEmptyImageIDs(ctx, matcher, d.deployment, d.images)
-	}
-	imageIDs := make([]string, 0, len(d.images))
-	for _, i := range d.images {
-		imageIDs = append(imageIDs, i.GetId())
-	}
-
-	violations, err := matcher.MatchOne(ctx, d.searcher, d.deployment.GetId(), &searchbasedpolicies.MatcherOpts{
-		IDFilters: map[v1.SearchCategory][]string{
-			v1.SearchCategory_IMAGES: imageIDs,
-		}})
-	if err != nil {
-		return nil, err
-	}
-	return violations.AlertViolations, nil
-}
-
-func matchWithEmptyImageIDs(ctx context.Context, matcher searchbasedpolicies.Matcher, deployment *storage.Deployment, images []*storage.Image) ([]*storage.Alert_Violation, error) {
-	closeableIndex, deploymentIndex, deploymentID, err := singleDeploymentSearcher(deployment, images)
+	closeableIndex, deploymentIndex, deploymentID, err := singleDeploymentSearcher(d.deployment, d.images)
 	if err != nil {
 		return nil, err
 	}